requirements.txt
networkx~=2.5.1
matplotlib~=3.4.1
SQLAlchemy~=1.4.5
opencv-python~=4.5.1.48
pandas~=1.2.3
pytesseract~=0.3.7
numpy~=1.20.2
PyGetWindow~=0.0.9
Pillow~=8.2.0
pynput~=1.7.3
logging~=0.4.9.6
psycopg2~=2.8.6